+++ src/platforms/Posix_Platform.cpp
@@ -43,7 +43,7 @@ namespace april
 	int64_t _getRamConsumption_platform()
 	{
 		hlog::warn(logTag, "Cannot use getRamConsumption() on this platform.");
-		return 0LL;
+		return (int64_t)0;
 	}
 	
 	void _getNotchOffsets_platform(gvec2i& topLeft, gvec2i& bottomRight, bool landscape)
